move UnhandledAction check
